Finally a working version to calculate the likelihood (fraction)
Events from beyond the maximum radius are shuffled uniform in
solid angle and appended to the real events.
This way, I can confirm that the maximum likelihood works and
can be used to estimate the fraction of signal and background
events.

Next step is to actually fit the number of signal events
instead of the fraction